% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_country_names.R
\name{check_country_names}
\alias{check_country_names}
\title{Check and convert country names to ISO3}
\usage{
check_country_names(countries, lookupfile)
}
\arguments{
\item{countries}{list of country character vercors to be checked from the data file}

\item{lookupfile}{the filepath of the lookup file}
}
\value{
If check failed returns "Failed". If check sucessfull returns ISO3 country codes
}
\description{
This function simply checks SOOS WG country data against the available lookup table. If the test is passed, it then converts SOOS country names into ISO3 codes for mapping.If the test is failed, a "Failed" message is returned.
}
\note{
In development
}
\author{
Kimberlee Baldry
}
