% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{}
\usage{
check_data(data, fun = "WG_maps")
}
\arguments{
\item{data}{The data frame that you are checking}

\item{fun}{Which data are you checking? One of "SOOSWG_file", "ISO3country_file", "geocode_file" or "category_file".}
}
\value{
It will let you know if the code worked!
}
\description{

}
\note{
In development. Editing.
}
\author{
Kimberlee Baldry
}
