% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_institution_names.R
\name{check_institution_names}
\alias{check_institution_names}
\title{Check and get geocodes of institutions}
\usage{
check_institution_names(institutions, institution_names_data)
}
\arguments{
\item{institutions}{list of institution character vercors to be checked from the data file}

\item{institution_names_data}{A data frame. The data frame must contain 4 columns labeled "Name" (what SOOS records the Affiliations as), "Type", "Latitude" and "Longitude". SSee example_institution_data for an example.#'}
}
\value{
If check failed returns "Failed". If check sucessfull returns institution,lat,lon as a datafram
}
\description{
This function simply checks SOOS institution data against the available geocode table. If the test is passed, it then retrieves lat, lon coordinates for SOOS institution names as a data frame. If the test is failed, a "Failed" message is returned.
}
\note{
v.0.1
}
\author{
Kimberlee Baldry
}
