% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_categories.R
\name{country_categories}
\alias{country_categories}
\title{Clean category data frame to help plot new country fill colours}
\usage{
country_categories(cat_data_orig, cat_countries_ISO3, cols, WG_i)
}
\arguments{
\item{cat_data_orig}{same as category_data in category_maps(). This data frame must contain 3 columns labeled "Acronym" (for the WG or plotname), "Countries" (A list of SOOS named countries separated by a ";") and "Categories" (a list of assigned categories for each SOOS country). See example file for details. Note the categories should be ordered from low-high and can be numbers or a caracter vector corresponding to names in the "Categories" column. The highest ranking category will be decide the country fill.}

\item{cat_countries_ISO3}{ISO3 codes returned by passing cat_data_original through check_country_names().}

\item{cols}{The ordered colour vector for categories.}

\item{WG_i}{The row of cat_data_orig that data is being generated for.}
}
\value{
a data frame with the columns "ISO3" (ISO3 ccountry code), "cat" (category ID) and "fill" (assigned fill colour)
}
\description{
This script contains a function to create a data frame to help reassign participating country fill in a ggplot object created from plot_WG_map() or plot_geocode_map().
}
\note{
v.0.1
}
\author{
Kimberlee Baldry
}
