% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_maps.R
\name{geocode_maps}
\alias{geocode_maps}
\title{Create multiple SOOS Geocoded Maps from a file}
\usage{
geocode_maps(
  SOOS_WG_data,
  outdir,
  country_names_data = NULL,
  institution_names_data,
  hilight_countries = F
)
}
\arguments{
\item{SOOS_WG_data}{A data frame. The data frame must contain the columns named "Acronym" (for the WG), "Affiliations" and "Countries.Represented" (optional). See "example_WG_data" for an example.}

\item{outdir}{The directory where you would like .png files saved}

\item{country_names_data}{A data frame. The data frame must contain 2 columns labeled "Country" (what SOOS records the country as) and "ISO3_name". See "example_country_data" for an example. This data frame is optional - do you want participating countries coloured?}

\item{institution_names_data}{A data frame. The data frame must contain 4 columns labeled "Name" (what SOOS records the Affiliations as), "Type", "Latitude" and "Longitude". SSee example_institution_data for an example.}

\item{add_countries}{If set to TRUE, participating countries will be coloured.}

\item{...}{see base_plot() and plot_WG_map() for further plot options (background_colour, world_map_colour, border_colour, plot_border_thickness)}
}
\value{
It will let you know if the code worked!
}
\description{
This script contains a function to create all SOOS WG maps from a file. The function: - acesses the file - applies data from rows of the file to 1. plot a WG map using plot_WG_map() and 2. save the file as a .png file using save_map() - outputs a message for each sucessfull map and when teh run is complete
}
\note{
In development. First Draft.
}
\author{
Kimberlee Baldry
}
