% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geocode_map.R
\name{plot_geocode_map}
\alias{plot_geocode_map}
\title{Draw a single SOOS Working Group Map with institution markers}
\usage{
plot_geocode_map(
  WG_name,
  WG_institutions,
  hilight_countries,
  WG_countries = NULL,
  participation_colour = SOOScol[1],
  country_border_colour = SOOScol[9],
  country_border_lwd = 1,
  marker_colour = NULL,
  marker_dot_colour = NULL,
  marker_size = 2.5,
  marker_asp = 1.1,
  marker_png = file.path(.libPaths()[1], "mapSOOSWG", "marker_base.png"),
  marker_dot_png = file.path(.libPaths()[1], "mapSOOSWG", "marker_dot.png"),
  add_marker = T,
  add_marker_dot = T,
  ...
)
}
\arguments{
\item{WG_name}{Name of the WG to be plotted}

\item{WG_institutions}{Dataframe of the institution geocodes to be plotted with column names as "Institution", "lat" and "lon"}

\item{hilight_countries}{If true, also colour the member countries.}

\item{WG_countries}{Character vector of the ISO3 codes of WG member countries}

\item{participation_colour}{The fill colour of the participating countries.}

\item{country_border_colour}{The colour of the participating country borders.}

\item{country_border_lwd}{The width of the participating country borders.}

\item{marker_colour}{The colour of the marker base, if not inherited from the PNG file. Leave as NULL if the PNG file contains the right colour. Note that no outline is possible if not drawn in the PNG file.}

\item{marker_dot_colour}{The colour of the marker dot, if not inherited from the PNG file. Leave as NULL if the PNG file contains the right colour.}

\item{marker_size}{The size of the marker.}

\item{marker_asp}{The apect ratio to adjust the marker.}

\item{marker_png}{The pathe to the base marker PNG file.}

\item{marker_dot_png}{The path to the marker dot PNG file.}

\item{add_marker}{If FALSE, no markers will be added.}

\item{add_marker_dot}{TRUE of a marker dot is being added to the plot.}

\item{...}{see base_plot() for further plot options (background_colour, world_map_colour, border_colour, plot_border_thickness)}
}
\value{
A ggplot object
}
\description{
This script contains a function to draw a map for a single working group with institution markers.
}
\note{
v.0.1
}
\author{
Kimberlee Baldry
}
