% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_WG_map_style1.R
\name{plot_WG_map_style1}
\alias{plot_WG_map_style1}
\title{Draw a single SOOS Working Group Map}
\usage{
plot_WG_map_style1(WG_name, WG_countries)
}
\arguments{
\item{WG_name}{Name of the WG to be plotted}

\item{WG_countries}{Character vector of the ISO3 codes of WG member countries}
}
\value{
A ggplot object
}
\description{
This script contains a function to draw a map for a single working group.
}
\note{
In development
}
\author{
Kimberlee Baldry
}
