% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_map.R
\name{save_map}
\alias{save_map}
\title{Save a single SOOSmap}
\usage{
save_map(plot, WG_name, outdir)
}
\arguments{
\item{plot}{A ggplot object}

\item{WG_name}{The name of the working group associated with the plot}

\item{outdir}{The directory where the file should be saved}
}
\description{
This script contains a function to save a ggplot map using ggsave as a PNG file
}
\note{
In development
}
\author{
Kimberlee Baldry
}
