% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_map.R
\name{save_map}
\alias{save_map}
\title{Save a single SOOSmap}
\usage{
save_map(plot, WG_name, outdir, p_width = 35, p_height = 25.8)
}
\arguments{
\item{plot}{A ggplot object}

\item{WG_name}{The name of the working group associated with the plot}

\item{outdir}{The directory where the file should be saved}

\item{p_width}{plot width in cm}

\item{p_height}{plot height in cm}
}
\description{
This script contains a function to save a ggplot map using ggsave as a PNG file
}
\note{
v.0.1
}
\author{
Kimberlee Baldry
}
