% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WG_maps.R
\name{WG_maps}
\alias{WG_maps}
\title{Create multiple SOOS Working Group Map from a file}
\usage{
WG_maps(filename, outdir, lookupfile)
}
\arguments{
\item{outdir}{The directory where you would like .png files saved}

\item{filepath}{The fullname of the file}
}
\value{
A ggplot object
}
\description{
This script contains a function to create all SOOS WG maps from a file. The function: - acesses the file - applies data from rows of the file to 1. plot a WG map using plot_WG_map() and 2. save the file as a .png file using save_map() - outputs a message for each sucessfull map and when teh run is complete
}
\note{
In development
}
\author{
Kimberlee Baldry
}
